quickdraw/requirements.txt
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-language==1.1.1
google-cloud-speech==0.36.3
googleapis-common-protos==1.56.2
google-cloud-translate
quickdraw==0.1.0
PyAudio